Add author conversion file for git conversion
Places an authors file in the new directory "gitconvert", which will
hold materials used in the svn->git conversion of the master.

Email addresses in the authors file are intended to represent the
address in use at the time the committer was active in the Kerberos
development community, even if that address doesn't work any more.  In
a few cases (austinj, edg, grier, shanzer), I wasn't able to find full
names.