Make rt-cvsgate accept commands at end of message
Modify find_commands to look for pseudo-headers at the end of a log
message as well as at the beginning.

The old version had the side effect of stripping empty lines at the
beginning of the message, but would leave a blank line at the start
if one followed the pseudo-headers.  The new version has weirder
side effects regarding blank line stripping; I expect it doesn't
really matter.