Update for Request Tracker 4.4.4
Update rt-cvsgate for RT 4 APIs.  Update krb5-daily.sh to dump the rt4
database and clean the RT 4 session data; for now it will continue to
run out of /var/rt2.

Remove send-pr-gate; we stopped shipping a krb5-send-pr in 1.13 and
haven't seen bug reports in its format in a while.  Remove
rt-mailgate, as we can use the stock version.

Add RT and Apache httpd configuration files to rt-scripts, both
updated for RT 4.