Update githooks for repository move
The authoritative repositories are moving from git.mit.edu to
krbdev.mit.edu.  Update the commit-msg hook for the new location and
convert the repository git hooks to Python 3.  The newer version of
git appears not to supply GIT_DIR to GIT_SSH so add a default value in
ssh-as-krbsnap.